requirements.txt
keyboard==0.13.5
Pillow==10.4.0
pystray==0.19.5
pywin32==306
tkinter_tooltip==3.1.0
